Model completed (two PurchaseNotice tests still broken)

+++ src/main/java/it/unical/asd/group6/computerSparePartsCompany/data/dao/ProductDao.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
-import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,15 +15,13 @@ public interface ProductDao extends JpaRepository<Product,Long>, JpaSpecificatio
 
     Optional<List<Product>> findAllByBrandAndModel(String brand, String model);
 
-    Optional<List<Product>> findAllByPrice(Integer price);
+    Optional<List<Product>> findAllByPrice(Double price);
 
-    Optional<List<Product>> findAllByPriceIsLessThan(Integer price);
+    Optional<List<Product>> findAllByPriceIsLessThan(Double price);
 
     Optional<List<Product>> findAllByModel(String model);
 
     Optional<List<Product>> findAllByCategory(String category);
 
     Optional<List<Product>> findAllByPriceBetween(Double p1, Double p2);
-
-    //FIXME QUERY DI ESEMPIO: CHIEDERE AD ANDREA E GIANFRANCO
 }
